+++ Source/ArrayExtension.swift
@@ -56,6 +56,13 @@ public extension Array {
         return Array(self[0..<Swift.max(0, Swift.min(number, count))])
     }
 
+    public func split(intoChunksOf chunkSize: Int) -> [[Element]] {
+        return stride(from: 0, to: self.count, by: chunkSize).map {
+            let endIndex = ($0.advanced(by: chunkSize) > self.count) ? self.count - $0 : chunkSize
+            return Array(self[$0..<$0.advanced(by: endIndex)])
+        }
+    }
+
 }
 
 public extension Array where Element : Equatable {